yield track segments from gpx parser to reduce memory footprint
The gpx parser is now a generator, yielding track segments as they
complete.  The resulting shapes will still need to be held in memory
all at once, but the source segments can be examined one by one and
released as we go.

A further improvement would be for the segment generator to yield
point generators so the segments could be cleared as we go.