Replace old Makefile with shell script running pytest
The old Makefile that ran tests needed to have them listed explicitly,
and it was easy to forget to add new tests to the list. Now we just run
pytest and let it discover everything.

And rather than have a Makefile with various targets, we just have a
shell script so we always run all tests and style checks.